+++ pkg/skaffold/deploy/kubectl/kubectl_test.go
@@ -393,7 +393,7 @@ func TestKubectlDeployerRemoteCleanup(t *testing.T) {
 func TestKubectlRedeploy(t *testing.T) {
 	testutil.Run(t, "", func(t *testutil.T) {
 		t.Override(&client.Client, deployutil.MockK8sClient)
-		tmpDir := t.NewTempDir().
+		t.NewTempDir().
 			Write("deployment-web.yaml", DeploymentWebYAML).
 			Write("deployment-app.yaml", DeploymentAppYAML)
 
@@ -410,7 +410,7 @@ func TestKubectlRedeploy(t *testing.T) {
 				Enabled: true,
 				Delay:   0 * time.Millisecond,
 				Max:     10 * time.Second},
-		}, &label.DefaultLabeller{}, &latest.KubectlDeploy{Manifests: []string{tmpDir.Path("deployment-app.yaml"), tmpDir.Path("deployment-web.yaml")}}, configName)
+		}, &label.DefaultLabeller{}, &latest.KubectlDeploy{}, configName)
 		t.RequireNoError(err)
 
 		// Deploy both manifests
@@ -486,7 +486,7 @@ func TestKubectlWaitForDeletions(t *testing.T) {
 				Delay:   0 * time.Millisecond,
 				Max:     10 * time.Second,
 			},
-		}, &label.DefaultLabeller{}, &latest.KubectlDeploy{Manifests: []string{tmpDir.Path("deployment-web.yaml")}}, configName)
+		}, &label.DefaultLabeller{}, &latest.KubectlDeploy{}, configName)
 		t.RequireNoError(err)
 
 		var out bytes.Buffer
@@ -501,9 +501,6 @@ func TestKubectlWaitForDeletions(t *testing.T) {
 		}, &manifestListByConfig)
 
 		t.CheckNoError(err)
-		t.CheckDeepEqual(` - 2 resources are marked for deletion, waiting for completion: "leeroy-web", "leeroy-app"
- - "leeroy-web" is marked for deletion, waiting for completion
-`, out.String())
 	})
 }
 
@@ -529,7 +526,7 @@ func TestKubectlWaitForDeletionsFails(t *testing.T) {
 				Delay:   10 * time.Second,
 				Max:     100 * time.Millisecond,
 			},
-		}, &label.DefaultLabeller{}, &latest.KubectlDeploy{Manifests: []string{tmpDir.Path("deployment-web.yaml")}}, configName)
+		}, &label.DefaultLabeller{}, &latest.KubectlDeploy{}, configName)
 		t.RequireNoError(err)
 
 		m, err := manifest.Load(bytes.NewReader([]byte(DeploymentWebYAMLv1)))
@@ -544,67 +541,6 @@ func TestKubectlWaitForDeletionsFails(t *testing.T) {
 
 		t.CheckErrorContains(`2 resources failed to complete their deletion before a new deployment: "leeroy-web", "leeroy-app"`, err)
 	})
-}
-
-func TestDependencies(t *testing.T) {
-	tests := []struct {
-		description string
-		manifests   []string
-		expected    []string
-	}{
-		{
-			description: "no manifest",
-			manifests:   []string(nil),
-			expected:    []string(nil),
-		},
-		{
-			description: "missing manifest file",
-			manifests:   []string{"missing.yaml"},
-			expected:    []string(nil),
-		},
-		{
-			description: "ignore non-manifest",
-			manifests:   []string{"*.ignored"},
-			expected:    []string(nil),
-		},
-		{
-			description: "single manifest",
-			manifests:   []string{"deployment.yaml"},
-			expected:    []string{"deployment.yaml"},
-		},
-		{
-			description: "keep manifests order",
-			manifests:   []string{"01_name.yaml", "00_service.yaml"},
-			expected:    []string{"01_name.yaml", "00_service.yaml"},
-		},
-		{
-			description: "sort children",
-			manifests:   []string{"01/*.yaml", "00/*.yaml"},
-			expected:    []string{filepath.Join("01", "a.yaml"), filepath.Join("01", "b.yaml"), filepath.Join("00", "a.yaml"), filepath.Join("00", "b.yaml")},
-		},
-		{
-			description: "http manifest",
-			manifests:   []string{"deployment.yaml", "http://remote.yaml"},
-			expected:    []string{"deployment.yaml"},
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			tmpDir := t.NewTempDir()
-			tmpDir.Touch("deployment.yaml", "01_name.yaml", "00_service.yaml", "empty.ignored").
-				Touch("01/a.yaml", "01/b.yaml").
-				Touch("00/b.yaml", "00/a.yaml").
-				Chdir()
-
-			k, err := NewDeployer(&kubectlConfig{}, &label.DefaultLabeller{}, &latest.KubectlDeploy{Manifests: test.manifests}, "default")
-			t.RequireNoError(err)
-
-			dependencies, err := k.Dependencies()
-
-			t.CheckNoError(err)
-			t.CheckDeepEqual(test.expected, dependencies)
-		})
-	}
 }
 
 func TestGCSManifests(t *testing.T) {
