+++ pkg/skaffold/render/renderer/kubectl/kubectl_test.go
@@ -18,6 +18,7 @@ package kubectl
 import (
 	"bytes"
 	"context"
+	"path/filepath"
 	"testing"
 
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/graph"
@@ -95,6 +96,75 @@ func TestRender(t *testing.T) {
 				true)
 			t.CheckNoError(errR)
 			t.CheckDeepEqual(test.expected, manifestList.String())
+		})
+	}
+}
+
+func TestDependencies(t *testing.T) {
+	tests := []struct {
+		description string
+		manifests   []string
+		expected    []string
+	}{
+		{
+			description: "no manifest",
+			manifests:   []string(nil),
+			expected:    []string(nil),
+		},
+		{
+			description: "missing manifest file",
+			manifests:   []string{"missing.yaml"},
+			expected:    []string(nil),
+		},
+		{
+			description: "ignore non-manifest",
+			manifests:   []string{"*.ignored"},
+			expected:    []string(nil),
+		},
+		{
+			description: "single manifest",
+			manifests:   []string{"deployment.yaml"},
+			expected:    []string{"deployment.yaml"},
+		},
+		{
+			description: "keep manifests order",
+			manifests:   []string{"01_name.yaml", "00_service.yaml"},
+			expected:    []string{"01_name.yaml", "00_service.yaml"},
+		},
+		{
+			description: "sort children",
+			manifests:   []string{"01/*.yaml", "00/*.yaml"},
+			expected:    []string{filepath.Join("01", "a.yaml"), filepath.Join("01", "b.yaml"), filepath.Join("00", "a.yaml"), filepath.Join("00", "b.yaml")},
+		},
+		{
+			description: "http manifest",
+			manifests:   []string{"deployment.yaml", "http://remote.yaml"},
+			expected:    []string{"deployment.yaml"},
+		},
+	}
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			tmpDir := t.NewTempDir()
+			tmpDir.Touch("deployment.yaml", "01_name.yaml", "00_service.yaml", "empty.ignored").
+				Touch("01/a.yaml", "01/b.yaml").
+				Touch("00/b.yaml", "00/a.yaml").
+				Chdir()
+
+			mockCfg := render.MockConfig{WorkingDir: tmpDir.Root()}
+			rCfg := latest.RenderConfig{
+				Generate: latest.Generate{RawK8s: test.manifests},
+			}
+			r, err := New(mockCfg, rCfg, map[string]string{}, "default")
+			t.CheckNoError(err)
+
+			dependencies, err := r.ManifestDeps()
+
+			t.CheckNoError(err)
+			expected := make([]string, len(test.expected))
+			for i, p := range test.expected {
+				expected[i] = filepath.Join(tmpDir.Root(), p)
+			}
+			t.CheckDeepEqual(expected, dependencies)
 		})
 	}
 }
